Update structeditor.App
Added ability to add the custom Dropdown "with actions" control
Added an options for supplying a cell array of numerics, which should be rendered as a space separated list of values in one input field (as opposed to a numeric array which can have 2 or 3 fields)
Added ability to show tips for each field (by clicking label)
Added option to add categorical arrays directly. Would have same effect as a config field with a cell array, i.e generate a popup/dropdown control
Added support for all numeric types.